Define *aoc_driver.service_names as const
In struct aoc_driver, change the definition of service_names such that
the pointer it points to is const and can't be modified. In other words,
service_names[n] should be const.

This helps with avoiding the follow checkpatch failure:

WARNING: static const char * array should probably be static const char * const

Change-Id: I2c468376da9441bc0701ddd61a85f8717565b05a
Signed-off-by: Daniel Mentz <danielmentz@google.com>