Pass cached address of DRAM carveout to subsystem coredump
The AoC driver maps the DRAM carveout as device memory, but the
subsystem coredump driver accesses the memory using normal memcpy().
This can sometimes cause a SError because the access is not the
native bus width. Instead, map the DRAM carveout as cached using
vmap() and pass that address to the subsystem coredump driver.

Bug: 160826318
Test: adb shell, su, echo "mem dump 0" > /dev/acd-debug, check no panic
Signed-off-by: Siqi Lin <siqilin@google.com>
Change-Id: I3961b1ea441594a3add44513ac97e1dedc4024f5