aoc: add support to get aoc ring base phys addr and ring size
To get the ring base physical address and the aoc ring size for pcm
MMAP. The physical address is inferred from the difference between
the ring base virtual address and the aoc_dram_virt_mapping. The
physical address of aoc_dram_virt_mapping is defined in the device tree
(dram_resource.start).

Bug: 163967290

Signed-off-by: Xinhui Zhou <zxinhui@google.com>
Change-Id: I8c3e0431a387220e7ff0386285cfe5ab2e04994c