Make access to physical address 0 non-fatal for ITMON
We're seeing some ITMON reports of AoC accessing physical address 0
with a transaction with the following attributes: SECURE, FIXED,
burst length 1, burst size 1 byte, read, error = DECERR.

Add an ITMON notifier and make the error non-fatal if the physical
address is 0 so the system can generate an AoC coredump. We can't
easily check the other transaction attributes because they require
parsing, so we only check for physical address = 0.

Bug: 174577569
Test: echo "mem dump 0xba400000" > /dev/acd-debug, AoC coredump generated
Test: echo "mem dump 0xba500000" > /dev/acd-debug, ITMON kernel panics
Signed-off-by: Siqi Lin <siqilin@google.com>
Change-Id: I43aaf6c7b9d3fc3beb5949c1d5c280a804c42eeb