aoc/alsa: add mix controls for sidetone EQ tuning
Added the following mixer controls for sidetone EQ tuning.

slider:/ # tinymix | grep Sidetone
119     BOOL    1       Sidetone Enable
120     INT     1       Sidetone Volume
121     INT     1       Sidetone Selected Mic
122     INT     1       Sidetone EQ Stage Number
123     INT     6       Sidetone Biquad0
124     INT     6       Sidetone Biquad1
125     INT     6       Sidetone Biquad2
126     INT     6       Sidetone Biquad3
127     INT     6       Sidetone Biquad4

There is no float type for mixer control value.
So for biquad parameters, input has to be integers obtained
by reinterpreting the 32 bit float type as int or unsigned int,

For example, for float point number 1, its 32 bit representeation
is 0x3f800000. It is 1065353216 if reinterpreted as a 32 bit int.

Bug: 176523095

Signed-off-by: Xinhui Zhou <zxinhui@google.com>
Change-Id: If4ac618d92402f3d63b2ec8eb42935ec8afb85af