Control AoC reset from AP
Previously, AoC firmware was in charge of resetting AoC by writing
the PCU_WATCHDOG_CONTROL register to generate an interrupt to
the APM. This reset can fail if AoC firmware crashes or hangs in
a really bad state, and it doesn't work if AoC crashes while AP
is in suspend. With this patch, AoC will generate an interrupt to
the AP, and the AP will ultimately reset AoC after collecting a
subsystem ramdump. It will also allow AP to reset AoC in response
to AoC hardware watchdog without any involvement from AoC firmware.

Bug: 178636702
Bug: 179696125
Change-Id: I3379039e6852f61f0b6c999b03cbe354ac8efb86
Signed-off-by: Siqi Lin <siqilin@google.com>