aoc/alsa: fix the kernel panic issue when aoc crashes
Disabling the set function in the mixer control for "PCM Playback
Volume', which has been used as a temporary solution for adjusting
playback volume. It became unnecessary now.

The AoC crash may lead to some alsa stream ptr not cleared properly.
As a result, when the mixer control set is called, the set function
tries to access the structure which is already freed.

There are other solutions to this issue (initalize structures in the
driver private data after AoC crash, or re-implement the set function
for volume control without using info from alsa streams).

But disabling the set function is most straightforward.

Bug: 179477638

Signed-off-by: Xinhui Zhou <zxinhui@google.com>
Change-Id: I796ff04401b2880f105422bf813ff525e4017a80