aoc/alsa: split mic capture start command for latency
Previously pdm mic is opened when the mic start cmd is executed.
But AoC will wait for 50 ms before sampling. This leads to issues
in audio capture latency and also missing interrupts for camera app
due to interrupt being disabled in snd_pcm_trigger. Now the 50 ms delay
is moved to the pcm prepare stage.

Bug: 178979913

Signed-off-by: Xinhui Zhou <zxinhui@google.com>
Change-Id: I349fe03afdffb238ff9a858879b4a1a805c49f08