Fix aoc_mbox_rx_callback kernel panic
There's a race condition when AoC sends a message immediately
after booting. The first "AoC booted" message causes the AoC
driver to set AoC to online and schedule a work item to
initialize the devices. When handling the second message, the
driver sees AoC is online and accesses the device structs while the
work item is still initializing them, causing a kernel panic
in wake_up() on the device read / write queues.

Bug: 182230878
Signed-off-by: Siqi Lin <siqilin@google.com>
Change-Id: I8a4efdb778119cf4d4f7eb06a84ebfab37fa4b64