aoc/alsa: advance write ptr for mmap playback on ULL
Support aaudio mmap playback on ULL by advancing the
write ptr artificially, so that AoC can advance the read ptr
accordingly.

It would be ideal for mmap playback if the AoC can update the
read ptr regardless of the write pointer position.

Raw playack uses mmap as aaudio mmap playback does. It also needs
this change if the pcm ack() is not called.

Bug: 179007309
Bug: 163967290

Signed-off-by: Xinhui Zhou <zxinhui@google.com>
Change-Id: I38315b7ec9f1f58335a4fe912a544fd1e06c8ba8