aoc: usb: avoid error accessing
If the headset is physical removed when we are doing
offload reset process, it may have null pointer access.
So we should check the XHCI status when we handle the irq.

Besides, use 100ms timeout for aoc service read/write in
order to avoid blocking in the function when AoC gets some
problem.

Bug: 184004173
Test: verify plug/unplug headset quickly
Signed-off-by: Puma Hsu <pumahsu@google.com>
Change-Id: Ie3b542ba321068f581f20363dfba87212de185a1