aoc: usb: config WQ_UNBOUND to alloc xhci workqueue
When the device works for video call, the CPU0/1 seems usually busy.
If an user inserts the headset at this moment, it might be delay to
handle xhci's irq work so that the System would hang for a long time.
Config WQ_UNBOUND so it will not bound to specific CPU.

Bug: 188004962
Test: cycle connecting headset during the video call.
Signed-off-by: Puma Hsu <pumahsu@google.com>
Change-Id: I4c129f5f676a1717628e8eb08e758441d7dbeb17