aoc/alsa: workaround to allow aoc service removal continue
aaudio service for mmap playback may duplicate multiple fd that
may cause issue in aoc service removal due to aoc crashing.

Currently, the removal process waits for the aoc service reference
number to be zero, but this could not happen in aaudio playback
due to mulitple duplicted fd, even though pcm_close() is called in HAL.

In this workaround, we do not wait for the reference number set to be
zero by pcm_close(). Instead it is done by pcm reset handler when aoc
is crashed.

Bug: 189031457
Bug: 187502463

Signed-off-by: Xinhui Zhou <zxinhui@google.com>
Change-Id: I36f9ab9f892be11d68b9539796dc1254c2e820ed