aoc/alsa: fix the deadlock issue for pcm substream lock
When aoc_audio_control() is called in the pcm triggering callback, the
interrupt is disabled. In this context, use spin_lock() instead of
spin_lock_bh() to avoid deadlock on pcm substream lock. The deadlock
was caused by spin_unlock_bh() which enables softirq processing.

See details on b/195165443#comment10

Bug: 195165443

Signed-off-by: Xinhui Zhou <zxinhui@google.com>
Change-Id: I69c4dea67b84bce5fcc5df7e5b25b3f958bc7bfe