aoc: usb: fix memory leakage in usb audio usecase
When the endpoint is not changed during allocating the bandwidth, xhci
driver free the old transfer ring and replace with the new one, the
leakage comes from setting the new ring to NULL when free the old ring.
Don't set the new ring to NULL in the condition to avoid memory leakage.

Bug: 196363789
Test: build and boot pass
Change-Id: I95532a84a0d285cbb55b55e9c5e0af3430698d16
Signed-off-by: Howard Yen <howardyen@google.com>