ANDROID: fix export made in "Revert "genirq: Restrict export of irq_to_desc()""
In commit 2cb5bf9e38e5 ("Revert "genirq: Restrict export of
irq_to_desc()"") the ability for the symbol to be exported was restored,
but the _GPL marking was removed for some reason.  As Android exports
always need to be _GPL, restore the _GPL marking of irq_to_desc() to
follow that rule, and to be synced with what is done upstream at the
same time.

Bug: 199823932
Fixes: 2cb5bf9e38e5 ("Revert "genirq: Restrict export of irq_to_desc()"")
Cc: Huang Yiwei <hyiwei@codeaurora.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>
Change-Id: Ib4e8d060805c8f59f9737abb77fc98e221e17a43
(cherry picked from commit df30793b7c7bef91f002dd48397c629de8c527dd)
Signed-off-by: Carlos Llamas <cmllamas@google.com>