pcie: exynos: Force msi_irq to -ENODEV
Even if MSI is enabled, we need to set msi_irq to -ENODEV.
This ensures that the core pci driver doesn't register the msi_irq in
dw_pcie_host_init(). The PCIe interrupt is registered during the probe
of this driver. The EP drivers register MSI interrupt handlers with the
core driver's MSI framework. When the EP has an MSI to send, it will
write to the MSI address specified in the PCIe config header which is
populated with an entry from the end point device tree.
The PCIe RC on detecting this write will generate an interrupt that is
routed to exynos_pcie_rc_irq_handler(). If it's an MSI interrupt, it
will be sent to the core pcie driver dw_handle_msi_irq(). This in turn
will route it to the appropriate handler registered by the EP driver.

Bug: 202338642
Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: Idc583562d06738b589856d665303d596ffca0527
Signed-off-by: Sajid Dalvi <sdalvi@google.com>