GKI: slider_gki: disable VH_I2C
This config is used to add special names to i2c devices. There are some
userspace dependencies on this in selinux and the pixel stats health
HAL, but from what I can tell the device is still functional without it.
So disable it for now and we can re-added it back if we run into any xTS
failures due to it being disabled.

Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: Ib8c35c4ca3e5f32007f64c960c627ac0294e6687