vendor_boot_modules.slider: rename ufs-pixel-fips140 module
The UFS fips module name was not added correctly to
vendor_boot_modules.slider, causing the module to be omitted in
initramfs. This caused a symbol to be missing when flashing images in
fastbootd.
This patch changes the name of the module in vendor_boot_modules.slider
from ufs-pixel-fips.ko to ufs-pixel-fips140.ko.

Bug: 185809932
Signed-off-by: Konstantin Vyshetsky <vkon@google.com>
Change-Id: I29093c82d782f46e9dd607b7a0a988e4098ff6bf