scsi: ufs: pixel: Fix dm-verity data corruption complaints
The SCSI core retries a command if deferred sense data is reported, if
the number of retries has not been exceeded and if none of the failfast
flags has been set. The block layer core sets the failfast flags for
readahead requests. Clear the failfast flags if the UFS device reports a
check condition to make the SCSI core retry the command.

Bug: 205394757
Change-Id: Ib861b7f04b53c3bee7999b3aaa5067ac6138fade
Signed-off-by: Bart Van Assche <bvanassche@google.com>